message-app-project/requirements.txt
Django==4.0
djangorestframework==3.13.1
requests==2.28.1